Moved Request/Response related code into different file
